+++ src/main/java/in/thirumal/persistence/model/Password.java
@@ -1,12 +1,10 @@
 package in.thirumal.persistence.model;
 
-import java.lang.Integer;
 import java.time.OffsetDateTime;
-import java.lang.Long;
-import java.lang.String;
+
 import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -28,7 +26,7 @@ public class Password implements java.io.Serializable  {
 
 	//Declaring fields
 	private Long passwordId;
-	private Integer loginId;
+	private Long loginId;
 	private String secret;
 	private OffsetDateTime startTime;
 	private OffsetDateTime endTime;
